working on inserting node+edges
