+++ helixdb/Cargo.toml
@@ -26,6 +26,8 @@ flume = "0.11.1"
 rayon = "1.8.0"
 polars = { version = "0.36", optional = true, features = ["parquet", "lazy"] }
 rusqlite = { version = "0.34", features = ["bundled"] }
+reqwest = { version = "0.11", features = ["json", "blocking"] }
+#serde_json = "1.0"
 
 [dev-dependencies]
 criterion = "0.5"
