+++ helixdb/src/helix_engine/mod.rs
@@ -3,4 +3,4 @@ pub mod graph_core;
 pub mod vector_core;
 pub mod types;
 pub mod macros;
-pub mod ingestion_engine;
+pub mod ingestion_core;
